pebble: fix compilation failure and atomic fields on 32bits
This commit addresses a few issues on 32bits including some
compilation failure because of int32 overflow and atomic fields
access. According to Go's issue: **On 32 bit platforms, only
64-bit aligned fields can be atomic**. So all atomic uint64/int64
fields should be aligned properly.

+++ options.go
@@ -917,7 +917,7 @@ func (o *Options) Validate() error {
 		fmt.Fprintf(&buf, "L0StopWritesThreshold (%d) must be >= L0CompactionThreshold (%d)\n",
 			o.L0StopWritesThreshold, o.L0CompactionThreshold)
 	}
-	if o.MemTableSize >= maxMemTableSize {
+	if uint64(o.MemTableSize) >= maxMemTableSize {
 		fmt.Fprintf(&buf, "MemTableSize (%s) must be < %s\n",
 			humanize.Uint64(uint64(o.MemTableSize)), humanize.Uint64(maxMemTableSize))
 	}
